+++ src/insight/index.ts
@@ -1,3 +1,1 @@
-export * from './insight.service';
-export * from './dtos';
-export * from './interfaces';
+export * from './responses';