+++ src/insight/interfaces/balance.response.ts
@@ -1,5 +1,0 @@
-export interface BalanceResponse {
-  user: string;
-  balance: number;
-  currency: string;
-}