+++ src/interfaces/termii.interface.ts
@@ -5,7 +5,6 @@ import { ModuleMetadata, Type } from '@nestjs/common';
  */
 export interface TermiiModuleOptions {
   apiKey: string;
-  senderId: string;
   baseUrl?: string;
 }