+++ src/messaging/dtos/index.ts
@@ -1,3 +1,0 @@
-export * from './send-message.dto';
-export * from './media.dto';
-export * from './request-sender-id.dto';