+++ src/messaging/interfaces/index.ts
@@ -1,5 +1,0 @@
-export * from './send-message.response';
-export * from './send-message.payload';
-export * from './request-sender-id.response';
-export * from './list-sender-ids.response';
-export * from './sender-id.interface';