+++ src/messaging/interfaces/list-sender-ids.response.ts
@@ -1,5 +1,0 @@
-import { SenderId } from './sender-id.interface';
-
-export interface ListSenderIdsResponse {
-  data: SenderId[];
-}