+++ src/messaging/interfaces/request-sender-id.response.ts
@@ -1,4 +1,0 @@
-export interface RequestSenderIdResponse {
-  code: string;
-  message: string;
-}