+++ src/messaging/interfaces/send-message.payload.ts
@@ -1,6 +1,0 @@
-import { SendMessageDto } from '../dtos';
-
-export interface TermiiSendMessagePayload extends SendMessageDto {
-  api_key: string;
-  from: string;
-}