+++ src/messaging/interfaces/send-message.response.ts
@@ -1,6 +1,0 @@
-export class SendMessageResponse {
-  message_id: string;
-  message: string;
-  balance: number;
-  user: string;
-}