+++ src/messaging/interfaces/sender-id.interface.ts
@@ -1,7 +1,0 @@
-export interface SenderId {
-  sender_id: string;
-  status: string;
-  usecase: string;
-  company: string;
-  created_at: string;
-}